Negative integers in @font-palette-values are invalid https://bugs.webkit.org/show_bug.cgi?id=230788 <rdar://problem/83528806>
Unreviewed post-review cleanup.

* css/parser/CSSParserImpl.cpp:
(WebCore::CSSParserImpl::consumeFontPaletteValuesRule):

Canonical link: https://commits.webkit.org/242505@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@283540 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/css/parser/CSSParserImpl.cpp
@@ -686,7 +686,7 @@ RefPtr<StyleRuleFontPaletteValues> CSSParserImpl::consumeFontPaletteValuesRule(C
         if (primitiveValue.isString())
             basePalette = FontPaletteIndex(primitiveValue.stringValue());
         else if (primitiveValue.isNumber())
-            basePalette = FontPaletteIndex(primitiveValue.value<int64_t>());
+            basePalette = FontPaletteIndex(primitiveValue.value<unsigned>());
         else if (primitiveValue.valueID() == CSSValueLight)
             basePalette = FontPaletteIndex(FontPaletteIndex::Type::Light);
         else if (primitiveValue.valueID() == CSSValueDark)
